pyproject.toml
[project]
name = "nov-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jahanzaib Tayyab", email = "jtayyab204@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.121.1",
    "openai-agents>=0.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[project.scripts]
nov-fastapi = "nov_fastapi:main"
nov-fastapi-prod = "nov_fastapi.main:main_prod"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
